Eclipse warnings cleanup
This change cleans up all warnings in Eclipse and adds Eclipse code formatter
settings.

The settings reflect what look like the general code style for the project.

Eclipse defaults with the following settings:
* Spaces not tabs, indent size of 4
* 120 line width for both code and comments

I noticed a lot of mixed tab/space formatting and whitespace, so this config
should probably be tweaked to reflect a agreed upon standard and all files
formatted to deal with the issues.

Pity m2e can't import this automatically:

https://issues.sonatype.org/browse/MNGECLIPSE-2240