Added maven shade config for titan-core
This builds a titan-core jar with the classifier "shaded03" containing
titan-core and all of its dependencies.  The shade plugin is
configured to prepend the string "titan03." to the package names of
all of the classes in this jar.

I also renamed the titan.properties file to titan03.properties to keep
it from being overwritten when the shaded03 jar is combined with a
titan-core jar from the 0.4.x or 0.5.y series.  In the future we need
to make this properties file package qualified,
i.e. com/thinkaurelius/titan/titan.properties, and we need to load it
in TitanConstants by calling getClass().getName().  This will allow
shading keep conflicting versions of the file from overwriting one
another.