Avoid serializing Guava Lists.transform() closure
HintedTitanGraph is serializable.  It assigned the return value of
Guava's Lists.transform(...) to a pair of final fields.  But the
returned list retained a reference to a Function, which in turn had
reference to a KeyRange.  KeyRange is not serializable, and all of
these retained references are implementation details that really
shouldn't be serialized in the first place.  Now HintedTitanGraph
takes care to eagerly evaluate the transform and store the result in a
standard-library ArrayList.