bin/ scripts to follow symlinks
Since many applications now symlink one, or more, times into multiple directories these
scripts needed updated to accurately follow said symlinks. Previously, if symlinks were
present, the script would fail to find all of the .jar files in the library because it
would only find the immediate path it was installed in. For example, if we symlinked
rexster.sh to /usr/local/bin/rexster and attempted to call it, the script would only
look in /usr/local/lib/ for its jar files which would fail to launch.