Rudimentary deprecation support in ConfigOption
This commit doesn't change the way ConfigOptions are read or written
to account for deprecation.  That's still the responsibility of client
code.  I didn't want to further complicate the config subsystem with
that, at least not yet.

This commit does add a new constructor that accepts a "deprecatedBy"
option of type ConfigOption, indicating that the option under
construction is deprecated in favor of the arrgument.
ConfigurationPrinter checks for this and formats deprecated options
differently; it draws their names with strikethrough and prepends a
red "Deprecated by..." warning to the struckthrough name.  Necessary
CSS and docbook tweaks also included in this commit.

This could be theoretically be extended so that BasicConfiguration
(and others?) are aware of deprecation and will favor the successor
option when it is set automatically, or to issue warnings.  I don't
want to tackle those quite yet.  Warnings are also harder than they
sound, since we don't want to constantly spam the log with them if a
certain piece of Titan repeatedly reads a deprecated option from its
config.
(cherry picked from commit b9aaad83ed4aa8028dd5e0a980812c49d76b869d)