Faunus Titan & graphson: drop negative vertex ids
This is a hack to deal with bogus graphson data containing negative
vertex ids. The negative ids are known to appear in _inE blocks and
might also appear elsewher, such as _outE or in _id.  This commit
attempts to make Faunus log and drop affected elements (the vertices
with negative ids and any edges incident on them).