Remove API doc links from reindex chapter
We pointed users of the TitanIndexRepair class to the API docs for
more information, but TitanIndexRepair is not included in the API docs
because those docs only cover titan-core.  Removing te links from the
chapter for now, but this is just a band-aid; TitanIndexRepair is a
user-facing public interface and should be part of the API docs in the
future.
(cherry picked from commit 967abf5410045c4c240fddc07b139c1a30078fed)