Add Solr troubleshooting entry on JTS jar
Solr can throw an exception during core creation when the jts jar is
absent from the Solr daemon classpath and schema.xml defines geo
fields.  This is a consequence of Spatial4j declaring JTS as an
optional dependency.

This commit adds a troubleshooting entry to the Solr docs that
summarizes the resulting exceptions and describes determination of the
appropriate JTS version to install.