Make Xdoclint:none activate only in JDK 1.8+
javadoc lint must be disabled on JDK 1.8 to prevent lint warnings from
killing the build, but JDK 1.7 doesn't understand the -Xdoclint:none
option that turns it off in 1.8.  JDK 1.7 responds to this option by
dying with nonzero status (which also kills the build).

This commit adds a Maven profile activated on JDK version range
"[1.8," that disables javadoc lint.