Update gh-pages release script
I'm modifying the old gh-pages-update.sh to work on the 0.5.x series.
The AsciiDoc changes between 0.4 and 0.5 made it unsafe to use on 0.5,
but now it just does javadoc and should be fine.  Also updated the
index.html template to match the current version live on the gh-pages
website right now.