Sleep in tests until ES is ready
ElasticsearchRunner.start now polls the ES logfile after starting the
daemon in the background until a line ending in " started" appears in
the log.  This change is aimed at CI, where the hard-coded 2 second
sleep wasn't long enough.  The method polls every 500 ms for up to one
minute, at which point it logs a message and returns.