Making HadoopPipeline.submit return Tool status
This submit method is a thin wrapper around ToolRunner.run.
ToolRunner.run returns the integer exit status produced by the Tool
implementation it executes.  Prior to this commit, submit returned
void and threw the status away.  Now the status is returned from
submit.  This change is groundwork for a new integration test method
related to #806 and #826 that will submit a HadoopPipeline that
submits a typical graphson-to-sequencefile dump job and needs to know
whether the job succeeded.