Move titan-dist test resources to test-classes
The titan-dist pom prior to this commit copied test resources to the
target/classes directory.  This doesn't get included in any of Titan's
release assemblies right now, but it's fragile.  Moved these resources
-- none of which should ever be included in a release archive -- to
target/test-classes to better reflect their function and reduce the
risk of accidental inclusion in future releases.