Moving TitanGraph instantiation into RecordReaders
This should address #817.  According to that issue, Spark (and maybe
MR?) creates an InputFormat when starting a workload just to get the
splits, then throws the IF away.  Creating a graph in IF.setConf
therefore lets an unclosed graph go out-of-scope in this scenario.

This commit implements @skavulya's suggestion to move graph
instantiation into RecordReader.initialize for both the HBase and
Cassandra RRs implementations.