Upgrading Avro version
This is part of #784.  The Avro version managed in the top-level pom
prior to this commit, 1.4.1, depended on paranamer, which in turn
depended on an unshaded copy of ASM 3.2:

[INFO] +- org.apache.hadoop:hadoop-client:jar:2.2.0:compile
[INFO] |  \- org.apache.hadoop:hadoop-common:jar:2.2.0:compile
[INFO] |     \- org.apache.avro:avro:jar:1.4.1:compile (version managed from 1.7.4)
[INFO] |        \- com.thoughtworks.paranamer:paranamer-ant:jar:2.2:compile
[INFO] |           \- com.thoughtworks.paranamer:paranamer-generator:jar:2.2:compile
[INFO] |              \- (asm:asm:jar:3.1:compile - version managed from 3.2; omitted for duplicate)

This commit updates to Avro 1.7.4, which depends on a different
paranamer version that apparently no longer requires unshaded ASM,
though I don't know whether that's because paranamer stopped using ASM
or whether they've got a shaded and repackaged copy of it somewhere in
their artifacts.