Attempt upgrade to Groovy 2.3.7
This is part of #784.  This commit overrides the Groovy version from
1.8.9 (what TP 2.5.0 compiled against and wants at runtime, but which
depends on unshaded ASM 3.x) to 2.3.7, which uses a shaded copy of ASM
built with the jarjar utility.  This is not ideal.  I recompiled
gremlin-groovy against 2.3.7 and ran the test suite, which passed, but
the Console class required some changes on account of the jline
upgrade (1.x to 2.x involved some backwards-incompatible changes to
jline's history package).  There could be complications here that I'm
not seeing yet.