Move shaded Groovy + ASM classes to external jar
This removes the shaded copy of Groovy 1.8.9 and ASM 3.2 introduced to
titan-core in bfcc99dbd71e90f4a2a17b24cf02ab484bf89396.  That worked
fine in the Maven build, tests, and in Rexster and gremlin.sh testing
from the zip archives (in other words, substantially everywhere), but
IDEA refused to compile the Groovy classes in this configuration.  The
Maven build worked; only IDEA failed to build.  It seemed to ignore
the shade plugin config in titan-core, manually compiling just
titan-core/src/main but leaving out the Groovy and ASM classes,
resulting in NoClassDefFound on groovy.lang.GroovyObject when building
the other modules.

It's possible to work around this IDEA limitation by manually adding
Groovy 1.8.9 as a global library or as a project library in all the
affected modules.  IDEA does build the project in that case.  But I
don't like that approach, since it means IDEA and Maven tests will use
different copies of the Groovy runtime.

In an attempt to placate IDEA and while trying to keep testing as
uniform as possible between Maven and IDEA, I pushed the shaded Groovy
+ ASM classes out to their own artifact now published to central,
named groovy-shaded-asm.  This moots IDEA's inability to comprehend
the shade plugin config in titan-core, since titan-core is no longer
responsible for shading (it's done in this new artifact).

This is all related to #784.