Add test that reads a list-valued prop with Faunus
This test is modeled on #807.  #807 is reproducible on 0.5.1 but has
been fixed in titan05 (0.5.2-SNAPSHOT) at the time of this commit.
This was probably fixed as part of earlier Faunus NPE issue fixing,
but I haven't bisected and don't know exactly which change did it.

This commit adds a test which would catch a regression on #807.