Fixing KyroSerializer to use Titan classloader
If Titan is running in a container (such as spark) and Kyro is in use by
the container (as is the case with spark), the classloader that Kyro
uses will be the parent classloader of the titan classloader. This means
that the classes titan uses will not be visible to it. This check-in
makes the Kyro implementation used by Titan use the same classloader as
Titan.