Disable Faunus type checks when guessing schema
FaunusElement does some cardinality/multiplicity integrity checks when
adding relations.  These are not meaningful when the
cardinality/multiplicity of a type is unspecified, such as when
reading from N-Triples and writing to SequenceFiles.

This commit disables the checks by default.  They can be reenabled
when writing to a Titan output format, though the checks are not a
complete data integrity guarantee when batch-loading is enabled.  In
particular, it's still possible to violate a uniqueness constraint on
a property key when batch-loading is enabled and the Titan-Hadoop
schema checks are enabled.