Support ThreadLocal and CHM for Kryo caching
This commit adds the `graph.kryo-instance-cache` option.  It takes
either the value THREAD_LOCAL (default) or CONCURRENT_HASH_MAP.  This
commit effectively restores the old Kryo caching behavior of
KryoSerializer, where it stored Kryos in a ThreadLocal, and where
those ThreadLocals are not necessarily removed when Titan shuts down.
The CHM setting provides an alternative that, unlike ThreadLocal, is
exposed to synchronization overhead when retrieving cached Kryos, but
which also releases all of its cached Kryo references when the graph
is shutdown.

For #840