Deduplicate common ES & Lucene support tests
LuceneIndexTest and ElasticSearchIndexTest both implement a
testSupport() method that runs assertions on index.supports(...) with
a variety of datatypes, mappings, and predicates.

Most (but not all) of these testSupport() method bodies were identical
between the two test classes.  The common segment has been extracted
to the shared parent class IndexProviderTest as a new method named
testCommonSupport().

This change is intended to simplify upcoming REGEX-related test
changes in LuceneIndexTest for #879