Modify LuceneIndex to support REGEX on STRINGs
Lucene (at least at our present version) does not seem to provide a
Filter implementation that does regex matching.  It does provide a
Query implementation for regex matching.  LuceneIndex was only built
to generate Filters; it always used MatchAllDocsQuery.  This change
modifies LuceneIndex to allow the convertQuery method to build both a
Filter and an accompanying Query.

The Lucene REGEX tests introduced in
8f712e70d767313fa4ad401bd295e9bff9fd97bf pass at the moment.

For #879