Tweak ConsistentKeyLockerTest's expired lock tests
This introduces a pair of tests around expired lock behavior:

* Check that our own expired locks throw a new ExpiredLockException
  class when checked

* Check that expired locks left behind by other instances, or by the
  current instance but with a different-than-expected write time, do
  not necessarily throw an exception in check (unless they prevent the
  transaction from verifying one or more of its own locks)

Previously, only the latter case was implemented and tested.  The
former case is essentially a special case for #892, wherein we want to
provide a more meaningful exception type and message in the event that
a transaction's locks expire while held.