Add "args" binding back into ScriptExecutor
The Titan ScriptExecutor used to set an "args" variable in the Groovy
Binding that pointed to the command-line argument list.  The Faunus
ScriptExecutor didn't do this.  I'm modifying the Faunus
ScriptExecutor to set "args" (in addition to the "a1", "a2", ... that
it already sets) for backwards-compatibility.  This could break
existing scripts that explicitly rely on "args" being absent from the
variable binding when the script starts, but that seems like an
extremely narrow class of scripts.  Scripts that first assign to
"args" and then do stuff with it should be unaffected.