Put a min of 100 on cache.tx-cache-size
This is a workaround for #893 that's likely to avoid race conditions
related to small tx-cache-size values in practice, though it does not
technically eliminate the underlying problem.

Prior to this commit, I investigated an alternative approach involving
WeakReference and a consuming ReferenceQueue drained from the same
thread that touches GuavaVertexCache, and that passed Titan's tests
appeared to log all the right things via slf4j, but it's still too
scary right to introduce so close to a release milestone.  The
assumption required to make that WeakReference approach work is that
dirty vertices are at all times reachable via at least one strong
reference *outside* GVC (on any GC root), but it's not completely
trivial to prove this for all of Titan's internals.  It's also
impossible to write a watertight test for the WeakReference approach
since its behavior is inextricably linked with GC.