Keep vertices together with Cassandra widerows
This commit adds a new TitanCassandraWideRecordReader class that gets
instantiated when titan-hadoop's configuration includes
cassandra.input.widerows=true.  This RR attempts to make as many calls
to the underlying widerows iterator as necessary to collect all
(column,value) entries before considering a FaunusVertex complete and
passing it up to the MapReduce task.

For #963