Use Groovysh `set` builtin for result peek sizes
When a gremlin.sh command evaluates to an iterable, the number of
initial elements in the iterable printed to the console is now
controlled by a pair of Groovysh settings:

* `set tp-olap-result-peek <int>` (default 50)
* `set tp-oltp-result-peek <int>` (default 500)

The OLAP setting controls the number of initial lines read out of
result files produced by executing a HadoopPipeline (i.e. a
titan-hadoop/Faunus job).

The OLTP setting controls the number of initial elements read out of
any other iterable result.

Negative ints are treated as if Integer.MAX_VALUE were specified.
Zero has slightly special interpretation; it completely disables
iterator peeking.  Zero means that the shell will not even call
hasNext to decide whether to print "...".  Positive ints are used
exactly as specified.

Settings are persistent across gremlin.sh sessions.  They're stored in
the Java user Preferences system under the node
"/org/codehaus/groovy/tools/shell".  On *NIX systems, this is usually
in ~/.java/.userPrefs/org/codehaus/groovy/tools/shell/prefs.xml.  You
may notice that the implementation in this commit does not actually
call Preferences.flush anywhere to guarantee persistence.  It relies
instead on this guarantee from the Preferences javadoc: "Normal
termination of the Java Virtual Machine will not result in the loss of
pending updates -- an explicit flush invocation is not required upon
termination to ensure that pending updates are made persistent."

For #924