Avoid stopwords in testContainsWithMultipleValues
2a99d97913d8c05c7a53ceca77af944452fcde4c brought a new test method,
testContainsWithMultipleValues, for #909.  This passed on solr, but I
failed to notice that it failed on lucene.

The test ran a CONTAINS query on "was here".  On Solr, the query
matches the expected property, but on Lucene, it doesn't match even
though the data are the same.

If I'm reading the Solr and Lucene docs and test configs correctly,
then I think "was" is configured as a stopword for both of them.  I
think I've also ruled out dynamic field mapping as a potential culprit
-- the field in the test is getting named 1|1_t, and there's a dynamic
mapping for *_t in the schema file.  The FieldType also looks right;
it's got to stopwords.txt.  I also see stopwords.txt pushed into the
Solr server config when I turn org.apache.solr logging up to DEBUG.
Really not sure where the discrepancy is.

Reconciling this difference was consuming a bunch of time for unclear
gains, so I eventually tapped out and just changed the query string
from "was here" to "here hercules".  The code exercised is the same as
when searching for "was here", but this avoids common stopwords.

For #965