Adding thin helper class for reindex on BDB
More generally, IndexJobs takes a StandardTitanGraph and runs an
IndexRepairJob (or ScanJob or VertexScanJob) over the open graph.
This API isn't final.  I've also got to properly dispose of the
Gremlin snippet (as a test and possibly as part of the asciidocs as a
code snippet, with some adaptation in each case) that I'm hastily
embedding as a comment at the bottom of IndexJobs.java.

For #756