Add hack to trick IDEA into using Kryo
IDEA doesn't understand Maven's shade plugin.  titan-core uses the shade
plugin on Kryo.  When IDEA tries to build titan-core, it seems to just
ignore the shade bits, and then we get linkage errors related to Kryo
when IDEA tries to execute tests using that bogus artifact that it just
built.

This profile reenables the unshaded Kryo dependency.  Not convinced this
is going to be tenable, since widening the existing gap between the
Maven and IDEA test execution environments can't be a good thing in the
long run.

The profile is disabled by default.