Modify StandardIDPool to allow failure recovery
Prior to this commit, an IDAuthority.getIDBlock call that threw an
exception or timed out would leave its StandardIDPool unable to
progress for the rest of its lifetime.  After the first
timeout/exception, StandardIDPool would never attempt another call to
IDAuthority.getIDBlock.  This commit changes StandardIDPool to let it
start a new call to IDAuthority.getIDBlock following an
exception/timeout.

This also fixes the test just introduced in
2e637dd69812a4a7eb879e9f34ae03a1cfbfbaac.

There is no backoff logic introduced in this commit.  The assumption
is that the IDAuthority implementation will contain a reasonable
amount of backoff logic (e.g. BackendOperation.execute).

For #992