Drop the TTL -1=immortal idea
The assumption that 0 TTL means no TTL has been set is already several
places in the code.  For instance, I've seen it as a constant
parameter given to a method that returns a TTL (used as the default
value) and as a default return in the else block of an if statement.
The code also does not respond well to negative or null values for at
least two reasons: Durations must be nonnegative, so any leak of -1
into a Duration implementation throws an exception, and nulls can
eventually lead to runtime exceptions when doing stuff like
element.value("~ttl").

I think -1 expresses "no TTL" more naturally than 0, but not by a wide
margin, and catching all the places where 0 is assumed looks quite
tricky.  Not worth it.