Avoid 'new Stopwatch'
Stopwatch.createStarted is the forward-compatible way to instantiate a
new Guava Stopwatch (as of 18).  I would normally consider the
converse problem -- backwards compat -- but StandardIDPool is already
using Stopwatch.createStarted on another line.  So, while it's true
that createStarted is only 15.0 and up, we already require 15 as a
minimum anyway in this file and probably others.