Assert that loaded key count is 2000 after commit
This test makes the assumption that the underlying store has the ability to read uncommitted transactions.
http://docs.oracle.com/cd/E17277_02/html/java/com/sleepycat/je/Cursor.html
This may very well be the behavior of Berkeley JE transactions as they are created in titan-berkeleyje, but it is not part of the OrderedKeyValueStore interface/specification. Actually, it seems like the isolation/lock modes are configurable in titan-berkeleyje.